refactor: combine two managers in one for login module
combine two managers in one for login module

MODULE-1-PART-3-gallery
